requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
python-multipart==0.0.9
openai==1.30.3
fpdf2==2.7.6
python-dotenv==1.0.1
pydantic==2.7.1
